feat: disallow connection to already connected port (#3)
closes #3
