+++ source/src/blueprint-deck/src/App.tsx
@@ -1,17 +1,16 @@
 import React from "react";
-import { Node } from "react-flow-renderer";
 import "./App.css";
-import { BlueprintDeck } from "./BlueprintDeck";
-import { NodeTypes } from "./createElements";
-import {BlueprintNodeData, NodeData} from "./NodeData";
-import { BaseNode } from "./nodes/BaseNode";
-import { TestDesign } from "./TestDesign";
-import { TestRegistry } from "./TestRegistry";
+import {BlueprintDeck} from "./BlueprintDeck";
+import {NodeTypes} from "./nodes/createElements";
+import {BlueprintNodeData} from "./NodeData";
+import {BaseNode} from "./nodes/BaseNode";
+import {TestDesign} from "./TestDesign";
+import {TestRegistry} from "./TestRegistry";
 
 const myTypes: NodeTypes = {
   TestNode: (node: BlueprintNodeData) => (
     <BaseNode node={node}>
-      <div>Hallo1</div>
+        TestMe
     </BaseNode>
   ),
 };
