--- source/src/blueprint-deck/src/nodes/checkConnectionValid.ts
@@ -1,3 +1,0 @@
-export const checkConnectionValid = () => {
-
-}
