+++ source/src/blueprint-deck/src/nodes/defaults/ActiveateNode.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+import {BlueprintNodeData} from "../../NodeData";
+import {BaseNode} from "../BaseNode";
+
+export const ActivateNode = (node: BlueprintNodeData) => {
+    return (
+        <BaseNode node={node} />
+    );
+};
